--- delivery-service/src/main/java/com/gaebal_easy/delivery/application/feign/HubLocationDto.java
@@ -1,16 +1,0 @@
-package com.gaebal_easy.delivery.application.feign;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class HubLocationDto {
-    private double latitude;
-    private double longitude;
-
-}
