--- delivery-service/src/main/java/com/gaebal_easy/delivery/application/service/KafkaProducerService.java
@@ -1,20 +1,0 @@
-//package com.gaebal_easy.delivery.application.service;
-//
-//import com.fasterxml.jackson.databind.ObjectMapper;
-//import lombok.RequiredArgsConstructor;
-//import lombok.extern.slf4j.Slf4j;
-//import org.springframework.kafka.core.KafkaTemplate;
-//import org.springframework.stereotype.Component;
-//
-//@Slf4j
-//@Component
-//@RequiredArgsConstructor
-//public class KafkaProducerService {
-//
-//	private static final String DELIVERY_ASSIGNED_COMPLETED_TOPIC = "delivery-assigned-topic";
-//
-//	private final KafkaTemplate<String, String> kafkaTemplate;
-//	private final ObjectMapper objectMapper;
-//
-//
-//}
