+++ delivery-service/src/main/java/com/gaebal_easy/delivery/domain/entity/DeliveryDetail.java
@@ -1,8 +1,6 @@
 package com.gaebal_easy.delivery.domain.entity;
 
-import com.gaebal_easy.delivery.application.dto.kafkaConsumerDto.KafkaRequireAddressToHubDto;
-import com.gaebal_easy.delivery.application.feign.HubDirectDto;
-import com.gaebal_easy.delivery.application.feign.HubRouteDto;
+import com.gaebal_easy.delivery.presentation.feign.HubDirectDto;
 import com.gaebal_easy.delivery.domain.enums.DeliveryStatus;
 import gaebal_easy.common.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
