+++ delivery-service/src/main/java/com/gaebal_easy/delivery/presentation/dto/kafkaConsumerDto/ProductRequestDto.java
@@ -1,0 +1,20 @@
+package com.gaebal_easy.delivery.presentation.dto.kafkaConsumerDto;
+
+import java.util.UUID;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductRequestDto {
+	private UUID productId;
+	private String productName;
+	private Long quantity;
+}
