+++ delivery-service/src/main/java/com/gaebal_easy/delivery/presentation/feign/HubDirectDto.java
@@ -1,0 +1,17 @@
+package com.gaebal_easy.delivery.presentation.feign;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+public class HubDirectDto {
+    private String departName;
+    private String arriveName;
+    private int time;
+    private double distance;
+}
