+++ delivery-service/src/main/java/com/gaebal_easy/delivery/presentation/feign/HubRouteDto.java
@@ -1,0 +1,22 @@
+package com.gaebal_easy.delivery.presentation.feign;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Builder
+@ToString
+public class HubRouteDto {
+    private String depart;
+    private String arrive;
+    private int totalRequiredTime;
+    private double totalDistance;
+    private List<String> visitHubName;
+}
