--- hub-service/src/main/java/com/gaebal_easy/client/hub/application/dto/kafkaConsumerDto/ProductRequestDto.java
@@ -1,16 +1,0 @@
-package com.gaebal_easy.client.hub.application.dto.kafkaConsumerDto;
-
-import lombok.*;
-
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductRequestDto {
-	private UUID productId;
-	private String productName;
-	private Long quantity;
-}
