--- hub-service/src/main/java/com/gaebal_easy/client/hub/application/dto/kafkaProducerDto/KafkaStoreCreateAssignDto.java
@@ -1,18 +1,0 @@
-package com.gaebal_easy.client.hub.application.dto.kafkaProducerDto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class KafkaStoreCreateAssignDto {
-    private UUID hubId;
-    private UUID storeId;
-}
-
-
-
-
