+++ hub-service/src/main/java/com/gaebal_easy/client/hub/presentation/dto/kafkaConsumerDto/KafkaOrderStoreInfoDto.java
@@ -1,0 +1,22 @@
+package com.gaebal_easy.client.hub.presentation.dto.kafkaConsumerDto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class KafkaOrderStoreInfoDto {
+    private UUID orderId;
+    private UUID productId;
+    private UUID supplyStoreHubId;
+    private String supplyStoreName;
+    private UUID receiptStoreHubId;
+    private String receiptStoreName;
+    private String receiptStoreAddress;
+    private String productName;
+    private long amount;
+}
