+++ hub-service/src/main/java/com/gaebal_easy/client/hub/presentation/dto/kafkaConsumerDto/KafkaStoreCreateDto.java
@@ -1,0 +1,15 @@
+package com.gaebal_easy.client.hub.presentation.dto.kafkaConsumerDto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class KafkaStoreCreateDto {
+    private UUID id;
+    private String address;
+}
